requirements.txt
Pillow
reportlib